requirements.txt
ccxt>=4.0
pandas>=2.0
numpy>=1.24
pandas-ta
python-dotenv>=1.0
apscheduler>=3.10
requests>=2.28
